refactor(eventstream): rename transformStreamController to transformStreamControllers
